Use val count of param with largest number of vals
Can’t use number of values for first param, since it may be empty (objectVar will have no value if it is a new resource). The number of params therefore also don’t need to match. Some unit tests not working in this commit and are ignored, to be fixed later.